added file submit system at reunion edition page and tweaked the reunion edition page

+++ reuniones/views.py
@@ -74,23 +74,59 @@ def reunionEdit(request, reunion_name):
     # reunionForm is pre-filled with old info so its easier to edit
     reunionForm = ReunionForm(initial={'my_name':reunion.name, 'my_date':reunion.date, 'my_description':reunion.description})
 
+    if request.method == 'GET':
+        #reset session so it can be used again without re-adding the same documents
+        request.session['documents'] = []
+
     
     # Submit ReunionForm with new values -> edition of basic Info
     if request.method == 'POST' and 'submitButton' in request.POST:
         reunionForm = ReunionForm(request.POST)
         if reunionForm.is_valid():
+
             reunion.name = reunionForm.cleaned_data['my_name']
             reunion.date = reunionForm.cleaned_data['my_date']
             reunion.description = reunionForm.cleaned_data['my_description']
             reunion.save()
+
+
             return redirect ('../../' + reunion.name + '/edit') 
     
     # Delete document buttons -> Uses AJAX
-    if request.method == 'POST' and request.headers.get('x-requested-with') == 'XMLHttpRequest':
+    if request.method == 'POST' and request.headers.get('x-requested-with') == 'XMLHttpRequest' and request.POST['post_type'] == 'file_deletion':
         docName = request.POST['documentName'].strip()
         document = Document.objects.filter(title = docName)
         document.delete()
         print(docName)
+
+    if request.method == 'POST' and request.headers.get('x-requested-with') == 'XMLHttpRequest' and request.POST['post_type'] == 'file_submit':
+        for document_id in request.session['documents']:
+            
+            # Get tempDocument with the file we want to add to the reunion
+            tempDocu = TempDocument.objects.get(id = document_id)
+            
+            # Duplicate the file of the tempDoc to media/documents (instead of temp)
+            # We need two separate files in order for there to not be any problems between the tempDocument and Document models
+
+            initial_path = tempDocu.file.path
+            file_name="documents/" + tempDocu.file.name[len("temp/documents/"):]
+            new_path = os.path.join(settings.MEDIA_ROOT,file_name)
+            shutil.copy(initial_path, new_path)
+            
+            # Open duplicated file in media/documents/ and create a new Document instance with the file linked to it
+            with open(initial_path, 'rb') as f:
+                file_obj = File(f)
+
+                new_document = Document()
+
+                new_document.title = tempDocu.title # Same title as the old document
+                new_document.file.save(os.path.basename(new_path), file_obj)
+                os.remove(new_path)
+                
+            # Add document to the reunion instance
+            reunion.documents.add(new_document)
+
+        print('simonsais')
 
     context = {
         'reunion':reunion,
@@ -210,5 +246,13 @@ def file_upload(request):
         # Save ID to session so it can then be used in the corresponding view
         request.session.setdefault('documents',[]).append(document.id)
         request.session.save()
-        print(request.session['documents'])
+    return JsonResponse({'post':'fasle'})
+
+def edition_file_upload(request, reunion_name):
+    if request.method == 'POST':
+        my_file=request.FILES.get('file')
+        document = TempDocument.objects.create(title=my_file.name, file=my_file, creation_date = timezone.now())
+        # Save ID to session so it can then be used in the corresponding view
+        request.session.setdefault('documents',[]).append(document.id)
+        request.session.save()
     return JsonResponse({'post':'fasle'})
